Feat(Shared) : handleLoad, useEffect 함수 Main, Header로 각각 분리
