+++ src/components/js/Nav.js
@@ -1,4 +1,6 @@
+import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
+import { getAccount } from "../../api/apiUrl";
 import logoImg from "../../Assets/logo.svg";
 import profileImg from "../../Assets/profile.svg";
 import "../css/Nav.css";
@@ -16,12 +18,25 @@ function Profile({ account }) {
   );
 }
 
-function Nav({ account }) {
+function Nav() {
+  const [account, setAccount] = useState({});
+
+  const handleLoad = async () => {
+    const nextAccount = await getAccount();
+    setAccount(nextAccount);
+  };
+
+  useEffect(() => {
+    handleLoad();
+  }, []);
+
   return (
     <div className="Nav_wrapper">
       <div className="Nav">
         <div className="Nav_left">
-          <a href="/"><img className="Nav_logo" src={logoImg} alt={logoImg}/></a>
+          <a href="/">
+            <img className="Nav_logo" src={logoImg} alt={logoImg} />
+          </a>
         </div>
         <div className="Nav_right">
           {account.email ? (
